add busn logic for all entities

+++ Employees-service/src/main/java/com/learningspringboot/samah/employees/repository/ResetVerificationTokenRepository.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface ResetVerificationTokenRepository extends JpaRepository<ResetVerificationToken, Long> {
     ResetVerificationToken findByToken(String token);
-    ResetVerificationToken findByUserId(Long id);
+    ResetVerificationToken findByUserId(Integer id);
 }
